requirements.txt
discord.py
PyNaCl
edge-tts
aiosqlite
